+++ changelog.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [2019.06]
-* Changed: Updated the Image Component and `the_tribe_image()` method to accept an image path
+* Changed: Updated the Image Component and `the_tribe_image()` method to accept an image URL fallback
 
 ## [2019.05]
 * Changed: Updated WordPress core to 5.1.1